requirements.txt
requests>=2.25.0
PyYAML>=6.0 